use proper offset type (int64, since max size is greater than 4gb for storage buffers in vk)
change tests to test properly for op codes

+++ source/slang/slang-ast-modifier.h
@@ -738,7 +738,7 @@ class GLSLOffsetLayoutAttribute : public Attribute
 {
     SLANG_AST_CLASS(GLSLOffsetLayoutAttribute)
 
-    int32_t offset;
+    int64_t offset;
 };
 
 class GLSLSimpleIntegerLayoutAttribute : public Attribute
